created wallet installation wizard
Implemented UserControls that are displaying different UIs on each step so no new form is opening. Wallet encryption is temporarily disabled, for this is only the skeleton and many functionalities have yet to be implemented. At this point, user can only create new seed and save it in a file to desktop.