users can now secure generated seed with password
users can now finally secure generated seed with password. When the application is closed, user can now choose the wallet or enter the path where the wallet is saved. If the file is encrypted, they will have to provide password to access it. If it is not encrypted, anybody can open it and use it.